+++ src/main/java/pl/betoncraft/betonquest/compatibility/citizens/CitizensHologram.java
@@ -218,16 +218,7 @@ public class CitizensHologram extends BukkitRunnable implements Listener {
                 }
 
                 for (final Player player : Bukkit.getOnlinePlayers()) {
-                    boolean visible = true;
-
-                    for (final ConditionID condition : npcHologram.config.conditions) {
-                        if (!BetonQuest.condition(PlayerConverter.getID(player), condition)) {
-                            visible = false;
-                            break;
-                        }
-                    }
-
-                    if (visible) {
+                    if (BetonQuest.conditions(PlayerConverter.getID(player), npcHologram.config.conditions)) {
                         hologramEnabled = true;
                         if (npcHologram.hologram == null) {
                             final Hologram hologram = HologramsAPI.createHologram(BetonQuest.getInstance(), npc.getStoredLocation().clone().add(npcHologram.config.vector));