+++ src/main/java/pl/betoncraft/betonquest/conditions/PartyCondition.java
@@ -1,5 +1,6 @@
 package pl.betoncraft.betonquest.conditions;
 
+import org.bukkit.Bukkit;
 import pl.betoncraft.betonquest.BetonQuest;
 import pl.betoncraft.betonquest.Instruction;
 import pl.betoncraft.betonquest.VariableNumber;
@@ -10,6 +11,8 @@ import pl.betoncraft.betonquest.id.ConditionID;
 import pl.betoncraft.betonquest.utils.Utils;
 
 import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.stream.Stream;
 
 /**
  * Checks the conditions for the whole party (including the player that started
@@ -26,9 +29,9 @@ public class PartyCondition extends Condition {
     public PartyCondition(final Instruction instruction) throws InstructionParseException {
         super(instruction, false);
         range = instruction.getVarNum();
-        conditions = instruction.getList(e -> instruction.getCondition(e)).toArray(new ConditionID[0]);
-        everyone = instruction.getList(instruction.getOptional("every"), e -> instruction.getCondition(e)).toArray(new ConditionID[0]);
-        anyone = instruction.getList(instruction.getOptional("any"), e -> instruction.getCondition(e)).toArray(new ConditionID[0]);
+        conditions = instruction.getList(instruction::getCondition).toArray(new ConditionID[0]);
+        everyone = instruction.getList(instruction.getOptional("every"), instruction::getCondition).toArray(new ConditionID[0]);
+        anyone = instruction.getList(instruction.getOptional("any"), instruction::getCondition).toArray(new ConditionID[0]);
         count = instruction.getVarNum(instruction.getOptional("count"));
     }
 
@@ -37,28 +40,20 @@ public class PartyCondition extends Condition {
         // get the party
         final ArrayList<String> members = Utils.getParty(playerID, range.getDouble(playerID), instruction.getPackage().getName(), conditions);
         // check every condition against every player - all of them must meet those conditions
-        for (final ConditionID condition : everyone) {
-            for (final String memberID : members) {
-                // if this condition wasn't met by someone, return false
-                if (!BetonQuest.condition(memberID, condition)) {
-                    return false;
-                }
-            }
+        final Stream<String> partyStream = Bukkit.isPrimaryThread() ? members.stream() : members.parallelStream();
+        if (!partyStream.allMatch(member -> BetonQuest.conditions(member, everyone))) {
+            return false;
         }
-        // check every condition against every player - at least one of them must meet each of those
-        for (final ConditionID condition : anyone) {
-            boolean met = false;
-            for (final String memberID : members) {
-                if (BetonQuest.condition(memberID, condition)) {
-                    met = true;
-                    break;
-                }
-            }
-            // if this condition wasn't met by anyone, return false
-            if (!met) {
-                return false;
-            }
+
+        // check every condition against every player - every condition must be met by at least one player
+        final Stream<ConditionID> anyoneStream = Bukkit.isPrimaryThread() ? Arrays.stream(anyone) : Arrays.stream(anyone).parallel();
+        if (!anyoneStream.allMatch(condition -> {
+            final Stream<String> memberStream = Bukkit.isPrimaryThread() ? members.stream() : members.parallelStream();
+            return memberStream.anyMatch(member -> BetonQuest.condition(member, condition));
+        })) {
+            return false;
         }
+
         // if the count is more than 0, we need to check if there are more
         // players in the party than required minimum
         final int pCount = count == null ? 0 : count.getInt(playerID);